requirements.txt
fastapi
uvicorn[standard]
mistralai
python-dotenv
pydantic
coverage
pytest-cov